+++ changelog.md
@@ -7,6 +7,8 @@
 * nested arrays are now possible
 * when cloning an object, dictionaries are now cloned as well.
 * when cloning models, models in lists are now cloned as well instead of shallow copied
+* prototype for IAsyncEnumerable for return of lists
+* correctly parse additional properties if additional properties is array, overwritten primitive type of combination of both
 
 ### 1.8.0
 * Allow to set the function authorization level